Move passive helpers and values so that the file structure is more logical
Move passive values to a new values file inside languageProcessing. Also move passive helpers to the right subfolders.
