+++ packages/yoastseo/spec/languageProcessing/values/ProminentWordSpec.js
