+++ packages/yoastseo/src/languageProcessing/index.js
@@ -6,7 +6,7 @@ import stripSpaces from "./helpers/sanitize/stripSpaces";
 import baseStemmer from "./helpers/morphology/baseStemmer";
 import getWords from "./helpers/word/getWords";
 import AbstractResearcher from "./AbstractResearcher";
-import splitSentence from "./helpers/passiveVoice/periphrastic/getSentencePartsSplitOnStopwords";
+import splitSentence from "./helpers/passiveVoice/periphrastic/freeAuxiliaryParticipleOrder/getSentencePartsSplitOnStopwords";
 import getPeriphrasticSentenceParts from "./helpers/passiveVoice/periphrastic/getSentenceParts";
 import determineSentencePartIsPassive from "./helpers/passiveVoice/periphrastic/determineSentencePartIsPassive";
 import flattenSortLength from "./helpers/morphology/flattenSortLength";
@@ -15,8 +15,8 @@ import buildFormRule from "./helpers/morphology/buildFormRule";
 import createRulesFromArrays, { createSingleRuleFromArray } from "./helpers/morphology/createRulesFromArrays";
 import matchRegularParticiples from "./helpers/passiveVoice/periphrastic/matchRegularParticiples";
 import checkException from "./helpers/passiveVoice/periphrastic/checkException.js";
-import directPrecedenceException from "./helpers/passiveVoice/directPrecedenceException";
-import precedenceException from "./helpers/passiveVoice/precedenceException";
+import directPrecedenceException from "./helpers/passiveVoice/periphrastic/directPrecedenceException";
+import precedenceException from "./helpers/passiveVoice/periphrastic/precedenceException";
 import nonDirectPrecedenceException from "./helpers/passiveVoice/periphrastic/freeAuxiliaryParticipleOrder/nonDirectParticiplePrecedenceException";
 import findMatchingEndingInArray from "./helpers/morphology/findMatchingEndingInArray";
 import * as regexHelpers from "./helpers/morphology/regexHelpers";
