+++ packages/yoastseo/src/languageProcessing/values/Clause.js
