+++ packages/yoastseo/src/languageProcessing/values/ProminentWord.js
